little context menu tidy up
